partial outline of repo struct and an add treecas feature.
Most of the gittree code already needs a refactor since it was all
read-only operations, and now we want some reading and some writing
(and some business logic about modes) and that all ought to be
happening with a single read pass.  so, todos there.

Little hung up trying to figure out what the expected behavior should
be if a user tries to add a single file, rather than a dir.  It's a
tricky question because with no dir, there's no place to store
properties or metadata (and while admittedly Gar doesn't preserve much
metadata, it's still not flat zero, so that's an eyebrow raiser).