correct-ish "add" feature.  Modulo one critical issue with distinguishing IO error codes...
which currently means it works the first time, and then dies the second time.

File operations return a sane and sensible AlreadyExists code; whether
it's directories or the rename function that's at issue, i'm not sure,
but the treecas commit is returning... uh... strange things.

also printf debugs remain, and the main method is a dummy.