more benchmarks, with data and notes, and one... wrong conclusion.
I started using more Criterion features to make a tiny bit DRYer code,
but I'm not very impressed with the grouping functions.  The textual
output they produce is still absolutely ghastly, and I've been munging
it with tee and grep in the shell.  It does not bring joy.
I would like to consider other benchmarking systems in the future.

But as to the wrong conclusion:

Shortly after writing all this, and still thinking it smells fishy,
I re-visited the documentation of io::copy, and realized I'm something
of a dolt: io::copy is *also* doing special things when it notices
it is file descriptors.  *That's* why it and fs::copy appear to be
neck-and-neck in all cases: they... are.

(Probably.  I haven't traced the whole set of internal indirections.
But they're both doing more than nothing special, that's for sure.)

So.  Neat.

I'm gonna keep this snapshot as a log, but correct the meditations
in the next commit.